+++ src/web/Renderer.ts
@@ -110,6 +110,7 @@ export type SpeechBubbleRenderingItem = BaseRenderingItem & {
     radius: number;
     offsetX: number;
     dimension: Dimension;
+    up: boolean;
 };
 
 export type TextRenderingItem = BaseRenderingItem & {
@@ -193,7 +194,7 @@ export class Renderer {
                             break;
                         case RenderingType.SPEECH_BUBBLE:
                             ctx.beginPath();
-                            ctx = roundRect(ctx, Math.round(item.position.x), Math.round(item.position.y), Math.round(item.dimension.width), Math.round(item.dimension.height), item.radius, item.relativeToScreen, Math.round(item.offsetX));
+                            ctx = roundRect(ctx, Math.round(item.position.x), Math.round(item.position.y), Math.round(item.dimension.width), Math.round(item.dimension.height), item.radius, item.up, Math.round(item.offsetX));
                             ctx.fillStyle = item.fillColor;
                             ctx.fill();
                             ctx.closePath();